requirements.txt
asgiref==3.4.1
channels==3.0.4
channels-redis==3.3.1
Deprecated==1.2.13
Django==3.2.9
django-rest-framework==0.1.0
djangorestframework==3.12.4
psycopg2==2.9.2
pytz==2021.3
redis==4.0.2
sqlparse==0.4.2
wrapt==1.13.3
